pyflaskazureblob/requirements.txt
flask
azure-storage-blob
flask_bcrypt
flask_login
email_validator